feature/check_login => 1. Added current-user.decorator. 2. Added auth.guard. 3.Added Current-user middleware. 4.Registered the current-user.middleware to user module.
Flow => Current-user middelware will get user data from session. If there is one then find the id in db and then decorate in req.currentUser. However, the request type which belongs to Express namespace does not has currentUser property. In order to make this happen, we have to merge declare atr the top of the file.
Then, AuthGuard get data from session userId which will be used later by other route. The if the request.session.userId does not present then the guard will prevent the client from access a route.

The CurrentUser.Decorator acts similar to @Params or @Body. It will capture request.currentUser which decorated from CurrentUser.Middleware and user as an parameter in route method.