Feature/check_login => Check login, Add current.user.decorator, current.user.middleware, authdataguard, authguard, /whoami and etc (#5)
* feature/check_login =>
1. Added current-user.decorator.
2. Added auth.guard.
3.Added Current-user middleware.
4.Registered the current-user.middleware to user module.

Flow => Current-user middelware will get user data from session. If there is one then find the id in db and then decorate in req.currentUser. However, the request type which belongs to Express namespace does not has currentUser property. In order to make this happen, we have to merge declare atr the top of the file.
Then, AuthGuard get data from session userId which will be used later by other route. The if the request.session.userId does not present then the guard will prevent the client from access a route.

The CurrentUser.Decorator acts similar to @Params or @Body. It will capture request.currentUser which decorated from CurrentUser.Middleware and user as an parameter in route method.

* feature/check_login =>
1. Added new method findOne to User.service so Current-user.middleware can use.
2. Added user/whoami route to user.controller so client can ask if the user has already logged in.

* feature/check_login => add AuthData guard to check if email password are presented.

* feature/check_login =>
1. Prevent /signin and /signup from process undefined email and password.
2. Enable cors allow only for a certain client to access with GET, POST,  PUT and DELETE.